+++ TDD_Matencio/BlastAisa.DigiBook/BlastAsia.DigiBook.Api/Controllers/EmployeesController.cs
@@ -30,33 +30,53 @@ namespace BlastAsia.DigiBook.Api.Controllers
         [HttpGet]
         public IActionResult GetEmployees(Guid? id)
         {
-            var result = new List<Employee>();
-            if(id == null)
-            {
-                result.AddRange(this.employeeRepository.Retreive());
-            }
-            else
-            {
-                var employee = this.employeeRepository.Retrieve(id.Value);
-                result.Add(employee);
-            }
+          
+                var result = new List<Employee>();
+                if (id == null)
+                {
+                    result.AddRange(this.employeeRepository.Retreive());
+                }
+                else
+                {
+                    var employee = this.employeeRepository.Retrieve(id.Value);
+                    result.Add(employee);
+                }
 
-            return Ok(result);
+                return Ok(result);
+           
         }
 
         [HttpPost]
         public IActionResult CreateEmployee(
             [FromBody] Employee employee)
         {
-            var result = this.employeeService.Save(Guid.Empty, employee);
+            try
+            {
+                if (employee == null)
+                {
+                    return BadRequest();
+                }
 
-            return CreatedAtAction("GetEmployees",
-                new { id = employee.employeeId }, result);
+                var result = this.employeeService.Save(Guid.Empty, employee);
+
+                return CreatedAtAction("GetContacts",
+                    new { id = employee.employeeId }, result);
+            }
+
+            catch (Exception)
+            {
+                return BadRequest();
+            }
         }
 
         [HttpDelete]
         public IActionResult DeleteEmployee(Guid id)
         {
+            var employeeToDelete = this.employeeRepository.Retrieve(id);
+            if (employeeToDelete == null)
+            {
+                return NotFound();
+            }
             this.employeeRepository.Delete(id);
             return NoContent();
         }
@@ -65,33 +85,55 @@ namespace BlastAsia.DigiBook.Api.Controllers
         public IActionResult UpdateEmployee(
             [FromBody] Employee employee, Guid id)
         {
-            var existingEmployee = employeeRepository.Retrieve(id);
+            try
+            {
+                if (employee == null)
+                {
+                    return BadRequest();
+                }
 
-            existingEmployee.ApplyChanges(employee);
+                var oldEmployee = this.employeeRepository.Retrieve(id);
+                if (oldEmployee == null)
+                {
+                    return NotFound();
+                }
 
-            this.employeeService.Save(id, employee);
+                oldEmployee.ApplyChanges(employee);
 
-            return Ok(employee);
+                var result = this.employeeService.Save(id, employee);
+
+                return Ok(oldEmployee);
+            }
+            catch (Exception)
+            {
+                return BadRequest();
+            }
         }
 
         [HttpPatch]
         public IActionResult PatchEmployee(
             [FromBody]JsonPatchDocument patchedEmployee, Guid id)
         {
-            if(patchedEmployee == null)
+            try
+            {
+                if (patchedEmployee == null)
+                {
+                    return BadRequest();
+                }
+                var employee = employeeRepository.Retrieve(id);
+                if (employee == null)
+                {
+                    return NotFound();
+                }
+                patchedEmployee.ApplyTo(employee);
+                employeeService.Save(id, employee);
+
+                return Ok(employee);
+            }
+            catch (Exception)
             {
                 return BadRequest();
             }
-            var employee = employeeRepository.Retrieve(id);
-            if(employee == null)
-            {
-                return NotFound();
-            }
-
-            patchedEmployee.ApplyTo(employee);
-            employeeService.Save(id, employee);
-
-            return Ok(employee);
         }
     }
 }
